use same function to create external charges in QMMM, ONIOM and THREE_LAYER

+++ src/energy_int_qmmm.h
@@ -378,7 +378,7 @@ namespace energy
         /**indizes of MM atoms*/
         std::vector<size_t> mm_indices;
         /**indizes of MM atoms that are taken into acoount for electrostatic interaction with QM region*/
-				std::vector<size_t> charge_indices;
+				std::vector<int> charge_indices;
 
         /**vector of length total number of atoms
         only those elements are filled whose position corresponds to QM atoms
